fix BIN_ROOT.  fix SKIN_ROOT.  test on debian and /home/weewx installations.

+++ bin/weecfg/extension.py
@@ -6,11 +6,12 @@
 #
 """Utilities for installing and removing extensions"""
 
-# The names of some root directories. For the extension pmon,
-#    -user/                          # The USER_ROOT subdirectory
-#    -user/installer/                # The EXT_ROOT subdirectory
-#    -user/installer/pmon/           # The extension's installer subdirectory
-#    -user/installer/pmon/install.py # The copy of the installer for the extension
+# As an example, here are the names of some reference directories for the
+# extension pmon (process monitor):
+#  -user/                          # The USER_ROOT subdirectory
+#  -user/installer/                # The EXT_ROOT subdirectory
+#  -user/installer/pmon/           # The extension's installer subdirectory
+#  -user/installer/pmon/install.py # The copy of the installer for the extension
 
 import os
 import shutil
@@ -38,18 +39,24 @@ class ExtensionEngine(object):
                  dry_run=False, logger=None):
         """Initializer for ExtensionEngine. 
         
-        config_path: Path to the configuration file. (Something like /home/weewx/weewx.conf)
+        config_path: Path to the configuration file.  For example, something
+                     like /home/weewx/weewx.conf)
 
-        config_dict: The configuration dictionary (the contents of the file at config_path).
+        config_dict: The configuration dictionary, i.e., the contents of the
+                     file at config_path.
 
-        tmpdir: A temporary directory to be used for extracting tarballs and the like [Optional]
+        tmpdir: A temporary directory to be used for extracting tarballs and
+                the like [Optional]
 
-        bin_root: A path to the root of the weewx binary files (Something like /home/weewx/bin).
-        Optional. If not given, it will be determined from the location of this file.
+        bin_root: Path to the location of the weewx binary files.  For example,
+                  something like /home/weewx/bin. Optional. If not specified,
+                  it will be guessed based on the location of this file.
         
-        dry_run: If True, all the steps will be printed out, but nothing will actually be done.
+        dry_run: If True, all the steps will be printed out, but nothing will
+                 actually be done.
         
-        logger: An instance of weecfg.Logger. This will be used to print things to the console.        
+        logger: An instance of weecfg.Logger. This will be used to print
+                things to the console.        
         """
         self.config_path = config_path
         self.config_dict = config_dict
@@ -57,7 +64,8 @@ class ExtensionEngine(object):
         self.tmpdir = tmpdir or '/var/tmp'
         self.dry_run = dry_run
 
-        self.root_dict = weecfg.extract_roots(self.config_path, self.config_dict, bin_root)
+        self.root_dict = weecfg.extract_roots(self.config_path,
+                                              self.config_dict, bin_root)
         self.logger.log("root dictionary: %s" % self.root_dict, 4)
         
     def enumerate_extensions(self):
@@ -76,12 +84,14 @@ class ExtensionEngine(object):
             self.logger.log("No extensions installed", level=0)
             
     def install_extension(self, extension_path):
-        """Install the extension that can be found at a given path."""
+        """Install the extension from extension_path, which may be a file or
+        directory"""
         self.logger.log("Request to install '%s'" % extension_path)
         if os.path.isfile(extension_path):
             # It's a file, hopefully a tarball. Extract it, then install
             try:
-                member_names = weecfg.extract_tarball(extension_path, self.tmpdir, self.logger)
+                member_names = weecfg.extract_tarball(extension_path,
+                                                      self.tmpdir, self.logger)
                 extension_reldir = os.path.commonprefix(member_names)
                 if extension_reldir == '':
                     raise InstallError("No common path in tarfile '%s'. Unable to install." % extension_path)
@@ -90,42 +100,47 @@ class ExtensionEngine(object):
             finally:
                 shutil.rmtree(extension_dir, ignore_errors=True)
         elif os.path.isdir(extension_path):
-            # It's a directory, presumably containing the extension. Install directly
+            # It's a directory, presumably containing the extension components.
+            # Install directly
             self.install_from_dir(extension_path)
         else:
-            raise InstallError("Extension '%s' not found or cannot be identified." % extension_path)
+            raise InstallError("Extension '%s' not found." % extension_path)
     
         self.logger.log("Finished installing extension '%s'" % extension_path)
 
     def install_from_dir(self, extension_dir):
-        """Install the extension that can be found in a given directory"""
-        self.logger.log("Request to install extension found in directory %s" % extension_dir, level=2)
+        """Install the extension whose components are in extension_dir"""
+        self.logger.log("Request to install extension found in directory %s" %
+                        extension_dir, level=2)
 
-        # The "installer" is actually a dictionary containing what is to be installed and
-        # where. The "installer_path" is the path to the file containing that dictionary.        
+        # The "installer" is actually a dictionary containing what is to be
+        # installed and where. The "installer_path" is the path to the file
+        # containing that dictionary.        
         installer_path, installer = weecfg.get_extension_installer(extension_dir)
         extension_name = installer.get('name', 'Unknown')
         self.logger.log("Found extension with name '%s'" % extension_name, level=2)
 
-        # Go through all the files used by the extension. A "source tuple" is something
-        # like (bin, [user/myext.py, user/otherext.py]). The first element is the
-        # directory the files go in, the second element is a list of files to be put
-        # in that directory
-        self.logger.log("Copying new files.", level=2)
+        # Go through all the files used by the extension. A "source tuple" is
+        # something like (bin, [user/myext.py, user/otherext.py]). The first
+        # element is the directory the files go in, the second element is a
+        # list of files to be put in that directory
+        self.logger.log("Copying new files", level=2)
         N = 0
         for source_tuple in installer['files']:
-            # For each set of sources, check and see if it's a type we know about
+            # For each set of sources, see if it's a type we know about
             for directory in ExtensionEngine.target_dirs:
                 # This will be something like 'bin', or 'skins':
                 source_type = os.path.commonprefix((source_tuple[0], directory))
-                # If there is a match, source_type will be something other than an empty string:
+                # If there is a match, source_type will be something other
+                # than an empty string:
                 if source_type:
                     # This will be something like 'BIN_ROOT' or 'SKIN_ROOT':
                     root_type = ExtensionEngine.target_dirs[source_type]
                     # Now go through all the files of the source tuple
                     for install_file in source_tuple[1]:
                         source_path = os.path.join(extension_dir, install_file)
-                        destination_path = os.path.abspath(os.path.join(self.root_dict[root_type], '..', install_file))
+                        dst_file = ExtensionEngine._strip_leading_dir(install_file)
+                        destination_path = os.path.abspath(os.path.join(self.root_dict[root_type], dst_file))
                         self.logger.log("Copying from '%s' to '%s'" % (source_path, destination_path), level=3)
                         if not self.dry_run:
                             try:
@@ -144,38 +159,47 @@ class ExtensionEngine(object):
         # Look for options that have to be injected into the configuration file
         if 'config' in installer:
             self.logger.log("Adding new sections to configuration file", level=2)
-            # Remember any new top-level sections (so we can inject a major comment block):
+            # Remember any new top-level sections (so we can inject a major
+            # comment block):
             for top_level in installer['config']:
                 if top_level not in self.config_dict:
                     new_top_level.append(top_level)
+                    self.logger.log("Adding section %s" % top_level, level=3)
 
             if not self.dry_run:
                 # Inject any new config data into the configuration file
                 weecfg.conditional_merge(self.config_dict, installer['config'])
                 
-                # Now include the major comment block for any new top level sections
+                # Now include the major comment block for any new top level
+                # sections
                 for new_section in new_top_level:
                     self.config_dict.comments[new_section] = weecfg.major_comment_block + \
                                 ["# Options for extension '%s'" % extension_name]
                     
                 save_config = True
-        
-        # Go through all the possible service groups and see if the extension provides
-        # a new one
-        self.logger.log("Adding new services.", level=2)
+            self.logger.log("Merged extension settings into configuration file", level=3)
+
+        # Go through all the possible service groups and see if the extension
+        # provides a new one
+        self.logger.log("Adding services to service lists", level=2)
         for service_group in all_service_groups:
             if service_group in installer:
                 extension_svcs = weeutil.weeutil.option_as_list(installer[service_group])
+                # Be sure that the leaf node is actually a list
+                svc_list = weeutil.weeutil.option_as_list(self.config_dict['Engine']['Services'][service_group])
                 for svc in extension_svcs:
                     # See if this service is already in the service group
-                    if svc not in self.config_dict['Engine']['Services'][service_group]:
+                    if svc not in svc_list:
                         if not self.dry_run:
-                            # Add the new service into the appropriate service group
-                            self.config_dict['Engine']['Services'][service_group].append(svc)
+                            # Add the new service into the appropriate
+                            # service group
+                            svc_list.append(svc)
+                            self.config_dict['Engine']['Services'][service_group] = svc_list
                             save_config = True
-                        self.logger.log("Added new service %s to %s" %(svc, service_group))
+                        self.logger.log("Added new service %s to %s" %(svc, service_group), level=3)
 
-        # Save the extension's install.py file in the extension's installer directory
+        # Save the extension's install.py file in the extension's installer
+        # directory for later use enumerating and uninstalling
         extension_installer_dir = os.path.join(self.root_dict['EXT_ROOT'], extension_name)
         self.logger.log("Saving installer file to %s" % extension_installer_dir)
         if not self.dry_run:
@@ -237,23 +261,33 @@ class ExtensionEngine(object):
         self.logger.log("Removing files.", level=2)
         N = 0
         for source_tuple in installer['files']:
-            # For each set of sources, check and see if it's a type we know about
+            # For each set of sources, see if it's a type we know about
             for directory in ExtensionEngine.target_dirs:
                 # This will be something like 'bin', or 'skins':
                 source_type = os.path.commonprefix((source_tuple[0], directory))
-                # If there is a match, source_type will be something other than an empty string:
+                # If there is a match, source_type will be something other
+                # than an empty string:
                 if source_type:
                     # This will be something like 'BIN_ROOT' or 'SKIN_ROOT':
                     root_type = ExtensionEngine.target_dirs[source_type]
                     # Now go through all the files of the source tuple
                     for install_file in source_tuple[1]:
-                        destination_path = os.path.abspath(os.path.join(self.root_dict[root_type], '..', install_file))
+                        dst_file = ExtensionEngine._strip_leading_dir(install_file)                     
+                        destination_path = os.path.abspath(os.path.join(self.root_dict[root_type], dst_file))
                         N += self.delete_file(destination_path)
-                        N += self.delete_file(destination_path.replace('.py', 'pyc'), False)
-                        N += self.delete_file(destination_path.replace('.py', 'pyo'), False)
-                    # If the directory is empty, delete it
-                    directory = os.path.abspath(os.path.join(self.root_dict[root_type], '..', source_tuple[0]))
-                    self.delete_directory(directory)
+                        if destination_path.endswith('.py'):
+                            N += self.delete_file(
+                                destination_path.replace('.py', '.pyc'), False)
+                            N += self.delete_file(
+                                destination_path.replace('.py', '.pyo'), False)
+                    # For skin subdirectories, if the directory is empty then
+                    # delete it.  This applies only to extension-specific
+                    # sub-directories.
+                    if root_type == 'SKIN_ROOT':
+                        dst_dir = ExtensionEngine._strip_leading_dir(source_tuple[0])
+                        directory = os.path.abspath(os.path.join(
+                                self. root_dict[root_type], dst_dir))
+                        self.delete_directory(directory)
                     break
             else:
                 sys.exit("Unknown destination for file %s" % source_tuple)
@@ -294,5 +328,11 @@ class ExtensionEngine(object):
                     shutil.rmtree(directory)
         except OSError, e:
             if report_errors:
-                self.logger.log("Delete failed on directory '%s': %s" % (directory, e), level=2)
-        
+                self.logger.log("Delete failed on directory '%s': %s" %
+                                (directory, e), level=2)
+
+    @staticmethod
+    def _strip_leading_dir(path):
+        idx = path.find('/')
+        if idx >= 0:
+            return path[idx + 1:]
